Add a pbxproj parser.
This change adds decodable models for the pbxproj parser, and honestly
it's so painful BUT we're now able to decode a pbxproj to a set of
structures representing the underlying plist.

The idea is to use this in a upcoming feature to better understand the
build settings and structures of the project when moving IR around to
target folders.