Ignore bundles when looking for targets.
Cocoapods, when building a pod library, will create an additional native
target (as a .bundle) for assets. This bundle takes on the same name as
the library it's being embedded into, which causes a clash in names in
the target dictionary used by `XcodeProject` to make targets to their
internal representation.

Firstly, ignore bundles if present. Secondly, emit an error warning
users when a potential target name clash occurs, and ask them to report
it for further investigation.

I wish Apple would document this format...